新建评测方案页面数据集表格添加及动态统计表格
